+++ ft_printf.h
@@ -21,7 +21,7 @@
 int		ft_printf(const char *format, ...);
 int		ft_print_char(int c);
 int		ft_print_str(char *str);
-int		ft_print_ptr(unsigned long long ptr);
+int		ft_print_ptr(unsigned long ptr);
 int		ft_print_nbr(int n);
 int		ft_print_unsigned(unsigned int n);
 int		ft_print_hex(unsigned int num, const char format);
